+++ Scripts/Fish.gd
@@ -31,7 +31,7 @@ func _physics_process(delta):
             if get_parent().get_node('Player').item_magnet_enabled:
                 if distance < 250:
                     var target_direction = (get_parent().get_node("Player").global_position - global_position).normalized();
-                    velocity = target_direction * 1000
+                    velocity = target_direction * ( get_parent().get_node("Player").speed + 200 )
             
         DESPAWNING:
             if $StateTimer.time_left == 0: